Crypto plugin 1st draft
